+++ resources/views/_components/expired-session.php
@@ -1,39 +1,4 @@
-<script>
-    $(function () {
-        const app = new App();
-
-        const login_modal = $("#login-modal");
-        const login_form = $("#login-form");
-
-        var is_session_expired = false;
-
-        function check_session() {
-            app.callAjax({
-                url: <?php echo json_encode($router->route('auth.expired')) ?>,
-                type: 'post',
-                success: function (response) {
-                    if(response.success) {
-                        login_modal.modal("show");
-                        is_session_expired = true;
-                    }
-                },
-                noLoad: true
-            });
-        }
-        
-        var count_interval = setInterval(function () {
-            check_session();
-            if(is_session_expired == true) {
-                clearInterval(count_interval);
-            }
-        }, 10000);
-
-        app.form(login_form, function (response) {
-            login_modal.modal('toggle');
-        });
-    });
-</script>
-
+<?php $v->insert('_scripts/expired-session.js'); ?>
 <div class="modal fade" tabindex="-1" role="dialog" aria-hidden="true" id="login-modal" 
     data-bs-backdrop="static" data-bs-keyboard="false">
     <div class="modal-dialog modal-md">