+++ resources/views/_scripts/notifications.js.php
@@ -1,0 +1,14 @@
+<script>
+    $(function () {
+        <?php if($dbNotifications): ?>
+        $("#bell-notifications").click(function () {
+            $.ajax({
+                url: <?php echo json_encode($router->route('user.notifications.markAllAsRead')) ?>,
+                type: 'patch',
+                data: {},
+                dataType: 'json'
+            });
+        });
+        <?php endif; ?>
+    });
+</script>