+++ resources/views/admin/users/_scripts/index.js.php
@@ -1,17 +1,16 @@
 <script>
     $(function () {
-        const app = new App();
         const table = $("#users");
         const filters_form = $("#filters");
 
-        const dataTable = app.table(table, table.data('action'));
-        dataTable.defaultParams(app.objectifyForm(filters_form)).filtersForm(filters_form)
-        .setMsgFunc((msg) => app.showMessage(msg.message, msg.type)).loadOnChange().addAction((table) => {
+        const dataTable = App.table(table, table.data('action'));
+        dataTable.defaultParams(App.form(filters_form).objectify()).filtersForm(filters_form)
+        .setMsgFunc((msg) => App.showMessage(msg.message, msg.type)).loadOnChange().addAction((table) => {
             table.find("[data-act=delete]").click(function () {
                 var data = $(this).data();
 
                 if(confirm(<?php echo json_encode(_('Deseja realmente excluir este usuário?')) ?>)) {
-                    app.callAjax({
+                    App.callAjax({
                         url: data.action,
                         type: data.method,
                         success: function (response) {