+++ docs/CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.9] - 2023-11-06
+### Added
+- DynamicForm class.
+
+## [1.7.8] - 2023-10-18
+### Added
+- Taking photos from the system through the Media Library.
+- FileSelector component.
+
 ## [1.7.7] - 2023-10-15
 ### Added
 - GTG MVC Core v1.0.10.